sumup/loyalty-analytics-platform/api/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
google-auth
pandas
python-multipart
pydantic
pydantic-settings
python-dateutil
pytz
numpy
